Use bpf_probe_read_user{,_str}
This makes sure we are accessing the target Ruby process' address space
and not some bogus kernel address. `bpf_probe_read` is no longer the
recommended option either, but we'll keep using the previous helper for
compatibility with older kernels.

Has some very naive way to detect which memory access helper we should
use. Maybe we could implement some smarter checks at runtime with a
sample programs that attempt to use the newer features we are interested
in.