Detect process id reuse
Process IDs (PIDs) are reused on Linux, so it's possible that we start profiling
a Ruby process with some PID and later on, while still profiling, this
process exits and a new one is spawned with this same PID.

The profiling information will be wrong. rbperf has some guardrails to
help here, such as ensuring that method and path names are valid
unicode, otherwise we consider the stack to be invalid, but it would be
best to actively ensure we don't have a race condition here.

By using the `pid + start_time` of the process, we should be able to have
a truly unique identifier of every process

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>