Add script to download + build libelf and zlib
usage:

```
$ tools/build_deps
$ export RUSTFLAGS="-L $PWD/target/static-libs"
$ cargo b # will now link against libelf and zlib
```

This allows us to have more control on the actual version of
both libraries that libbpf depends on, rather than always using
the ones installed by a package manager or a previous manual
installation.

They are compiled with clang and frame pointers. See comment
for the warnings that were disabled to compile elfutils with clang.

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>